pyproject.toml
[project]
name = "av1pack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.1.0",
    "tqdm>=4.67.1",
]
